Update FE - I don't know what i had done :v - Maybe Call API

+++ src/Server/Domain/DTO/Product/ProductEditModel.cs
@@ -11,8 +11,8 @@ namespace Domain.DTO.Product
         public string Description { get; set; }
         public string Specification { get; set; }
         public int Amount { get; set; }
-        public decimal Price { get; set; }
-        public decimal OrPrice { get; set; }
+        public int Price { get; set; }
+        public int OrPrice { get; set; }
         public int BranchId { get; set; }
         public int CategoryId { get; set; }
         public string TagSlug { get; set; }
